[Rebase to modules] EM/charger:  Fix MAX and MIN battery temperature settings
BZ: 118790

PMIC MAX and MIN temperature settings are done at charger driver
initialization, it is not required to handle it when Framework
provides the MAX and MIN value. This is due to PMIC providing a
fixed MAX and MIN ranges instead of each temperature zone defined
in the SBCT table.

Change-Id: I856ed57eecde4b61266a6b0e39f9ec0b3a2b636e
Signed-off-by: Naveen Rapaka <naveen.rapaka@intel.com>
Reviewed-on: http://android.intel.com:8080/117152
Reviewed-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>
Tested-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>