[Telephony/HSI_Trace]: Send OPEN_CONN eDLP cmd in open func
BZ: 138194

In order to allow the BPLOG trace to be evacuated over HSI
when the HSI is not used as IPC, the modem needs to receive
the OPEN_CONN request for channel 4 when AP is ready. Then
the mdm_trace_open function now requests the sending of
this eDLP command and completes when the ACK for this
command is received from the modem.

Change-Id: Iacb1ef7bf4fc3444cbe5492a93ffa5ec763b78b0
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>