[Telephony/HSI_TTY]: set tty->flow_stopped when MMGR closes TTYIFX0
BZ: 121401

In the TTY driver, in the closing procedure, the tty
link drivers are flushed if the flow_stopped flag is
set. Then the tty driver waits for the complete
remaining data to be transmitted if needed (this is
only applicable when the link is still working). In
the FMMO, the MMGR does not wait for the client
acknowledge to close the TTY_IFX0. So, when the MUX
driver tries to close its tty ports, the link is out
of order, but as the flow_stopped flag was not set,
it waits for the TX to be completed (no flush, so
it can have some data in the TX fifo) that never
finishes until the timer expires (up to 1 minute).

Change-Id: Ia2de1e0c4e08bfb073fe1c525ea807083fabc3d7
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>
Reviewed-on: http://android.intel.com:8080/120702
Reviewed-by: Daragnes, Julie <julie.daragnes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Noziska, Patrick J <patrick.j.noziska@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>