Reboot to COS upon decision made in userland during shutdown
BZ: 120649

The goal of this patch is to cleanup some code and to avoid
using power supply functions in OSIP driver for deciding to
reboot to COS during graceful shutdown.

Now the reboot to COS is made depending on a decision taken
at userland (Java).
If it is decided to reboot to COS at userland level, the OSIP
reboot notifier will be called with "charging" passed as
parameter.
With the new behavior some uneeded code can be removed now.

Change-Id: I2897d0c6b2396e66ab12f2a98306b31e72c3f7e7
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>
Reviewed-on: http://android.intel.com:8080/121826
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Muller, Francois-nicolas <francois-nicolas.muller@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Romieu, Benoit <benoit.romieu@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>