EM/Charger: Fix battery charging beyond 45 deg on RHB platform
BZ: 127286

RHB platform battery supports charging only till 45 deg. PMIC should
be programmed to disable charging if the battery temperature goes above
45 deg.
This patch ensures PMIC is programmed to charge the battery till 45 deg
and disable charging if the battery temperature goes above 45 deg.
This is a temporary change and logic needs to be devised to get the temperature
data from the battery profile.

Signed-off-by: Pandey Raj <raj.pandey@intel.com>

Change-Id: I8cb7f5b07bea0415fd47a034c9a2786926dbae8e
Reviewed-on: http://android.intel.com:8080/122125
Reviewed-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>
Tested-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>