[REVERTME] Fix OSNIB magic header corruption.
BZ: 127335

Standard kernel code is accessing in the zone we are using for OSNIB.
We will use a new address to avoid corruption of the OSNIB header.
Waiting for this (needs IAFW synchronization), here is a patch that allow
to detect both old and new address, and also is able to fix OSNIB header.

Signed-off-by: Guillaume Betous <guillaumex.betous@intel.com>

Change-Id: I47bf880bfa274c836809e4fcddf57bbb861805e2
Reviewed-on: http://android.intel.com:8080/123156
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>