TP2L: Tracepoint To Log - initial commit
BZ: 145657

TP2L stands for Tracepoint To Log.
TP2L module is responsible for doing the binding between the tracepoint
events (as defined in Documentation/trace/tracepoints.txt) and Intel's
proprietary trace infrastructure.

For each tracepoint event that is supported by TP2L, a new probe is created
to export the event data (ascii string defined by the event's TP_printk macro)
to the STM HW for an output either on the PTI port or the USB.DvC.Trace interface.

A debugfs is created (/sys/kernel/debug/tp2l) to handle the enabling of the
TP2L probe for each tracepoint event.

Change-Id: I267ce99c3a8f34c09f8fed2b6347331b504c5ff1
Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>