[Telephony/eDLP_Driver]: Check PDU offset is valid
BZ: 121844

If PDU signature is valid but offset is corrupted, the buffer
address computation leads to a wrong unmapped address and to
an Ipanic when trying to access it.

The offset range must be checked in order to prevent an illegal
access. In that case the PDU will be discarded.

Change-Id: Iadb472c1abca0e2642c8f30055f99a226e4eb928
Signed-off-by: Julie Daragnes <julie.daragnes@intel.com>
Reviewed-on: http://android.intel.com:8080/124202
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Reviewed-by: Bourree, Gwenn <gwenn.bourree@intel.com>
Reviewed-by: Kouassu, Alain <alain.kouassu@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Koskinen, Ilkka <ilkka.koskinen@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>