EM/pmic_ccsm: Hold a wake lock for 5s during USB ID detection
BZ: 132004

When the target is in Sleep (suspend to mem) state and if an
USB device (e.g. mouse, keypad or another phone) is conected
to the target through OTG cable, PMIC_CCSM will receive an
interrupt and it will duly notify the DWC OTG driver. But
since the OTG driver needs more time to process this
notification, after that notify the charger driver to enable
boost mode and enumerate the connected device etc, and Since
no wake locks are active, target again starts to enter
the suspend state before the OTG driver can successfully
enumerate the connected device.
To avoid this, PMIC-CCSM driver needs to hold the pmic
wake lock in pmic_isr if the USB_ID_DET or VBUS_DET bit is
set and it should release the wakelock after 5 secs. This
5secs will give enough time for the dwc_otg driver to
successfully enumerate the connected USB OTG device.
A new wakelock for bq24261 charger driver has been created
to prevent the system from going into suspend mode when a
charging cable has been connected to the target.

Change-Id: I025ce193ffabcd27c5f1f5f94508152edadecdc6
Signed-off-by: raosrini <srinidhi.rao@intel.com>
Reviewed-on: http://android.intel.com:8080/133591
Reviewed-by: Dower, Robert <robert.dower@intel.com>
Tested-by: Dower, Robert <robert.dower@intel.com>