wifi: avoid returning error on gpio_set_debounce().
BZ: 127461

gpio_set_debounce() is not supported on BYT. as a result, wifi fails
during the probe. This patch fixes the issue by removing the return when
gpio_set_debounce() fails.

Please note that gpio_set_debounce() is a workaround to a hardware issue on
Clovertrail SoC. It is not needed anymore on Saltbay and BYT. It can be also
removed on Clovertrail if the platform is using fast IRQ configured on level
trigger.

Change-Id: I9cf683a0ae7575ee2221a688611f0de4b2d4c293
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/127192
Reviewed-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>