vibra: Configure Vibra Enable GPIO in probe
BZ: 136937

Baytrail vibra uses PMIC GPIO for vibra enable/disable. This
GPIO is to be configured with drive type as CMOS and pull disable.
This configuration was earlier done in PWM configure function.
Since this configuration is a onetime operation, moved it to the
vibra probe instead.

Also changed the GPIO configuration value to not enable the
GPIO by default.Enable/disable will be done by the standard
GPIO api: gpio_set_value()

Change-Id: Ibb73a5784fd8d51ce582a2ddacd9ab527ae7d0d5
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>
Reviewed-on: http://android.intel.com:8080/131608
Reviewed-by: Shah, Hardik T <hardik.t.shah@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Abdullah, Omair M <omair.m.abdullah@intel.com>
Reviewed-by: Amireddy, MallikarjunareddyX <mallikarjunareddyx.amireddy@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Reviewed-by: Nilofer, Samreen <samreen.nilofer@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>