[Telephony/HSI-eDLP]: add scalability
BZ: 138194

In order to adapt the Trace channel opening to the modem,
the scalability has to be implemented to get the modem
and cpu version.
Once this step is complete, the Trace opening mechanism
uses the modem version to select the right way to use
the eDLP protocol and open the Trace channel.

Change-Id: Ibd6071d93164728cbeede3ed109db065bb70e041
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>