[PORT FROM K3.4] bcm: Enable fast IRQ for saltbay products
BZ: 118722

Define the HW_OOB CFLAG required so that the driver is using level
configuration for interrupts. This should be merged with an IFWI supporting
fast IRQ for "WLAN-interrupt" as exported by the SFI table, directly routed to
the IO-APIC rather than through GPIO controller.

This patch (re)disables interrupts in case of Level interrupt in ISR, as
pointed out by Frode Isaksen, so that we are not stormed by interrupts until
it's properly handled by dhd thread.

Also need to set the flag IRQF_ONESHOT so that IO-APIC interrupts are masked
while in the ISR.  Without this flag, there are twice as many interrupts.

Change-Id: I1403e1c33caba568df6297e0ce4239a8b9955c1d
Orig-Change-Id: I19b91ef361c212a64a79d41c8af5212d9f94c5e7
Signed-off-by: Quentin Casasnovas <quentinx.casasnovas@intel.com>
Reviewed-on: http://android.intel.com:8080/135578
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Ernst, Eric <eric.ernst@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>