[PORT FROM R43][Telephony/HSI]Add hsi controller suspend/resume notification
BZ: 122568

The platform exits S3. CPU0 is waken up and executes the resume
sequence. The TCP/IP retransmit timer fires and preempts the
resume worker. dlp_net_start_xmit is executed in the SW IRQ
context of the timer and tries to program a HSI transfer.
As the HSI resume callback hasn't been invoked yet, the device
is not ready and hsi_transfer function exits without programming
anything and without any error. Then the HSI TX timer is enabled and
expires before the resume is invoked.
This patch allows the HSI controller to notify HSI clients of
power management transitions(suspend/resume).When the suspend
event is received by the client(dlp_net in our case) this one
stops any transfer request from the IP stack and not restart it
until the resume event is received with this behavior we ensure
that a transfer will not be performed when the HW is not totally
resumed.

Change-Id: I1d2052e052ac08a8a0b7b06b91508a734e7985a2
Orig-Change-Id: I4b80071da98c1046688d85b348a9f400e83e868d
Signed-off-by: Mustapha Ben Zoubeir <mustaphax.ben.zoubeir@intel.com>
Signed-off-by: Julie Daragnes <julie.daragnes@intel.com>