[PORT FROM R4.3-Stable]EM/Charger: Fix for disabling the charging when the invalid battery is connected.
BZ: 137724

Issue Description:
Battery is charging when the invalid battery is connected,
and UI is showing "not charging"

Solution:
Disable the charging in bq24192_enable_charging()
when the invalid battery is connected. This was not handled
earlier in driver when framwork sent charging enable false
for invalid battery.

Category: device enablement
Domain: batteryManagement-TI
Origin: internal
Upstream-Candidate: no, needs rework

Origin-Change-Id: Id6757715582eb26eebf8f89a2df21947b8e4cf70
Signed-off-by: Vineesh KK <vineesh.k.k@intel.com>

Change-Id: I7ebaff7449a491dc31b0a720f12003158a6a69d9
Signed-off-by: Siddalingesh Koosalapur <siddalingesh.s.koosalapur@intel.com>