[EM/BCU]: Changes to BCU driver to avoid unwanted interrupts.
BZ: 146529

Issue Description: For BYT ffrd8, we observe too many bcu interrupts
for VWARNA and VWARNB with quick changes in vsys level. Also on unmasking
VWARNA we do not get pending interrupts that have
occurred when the VWARNA is masked.

Solution: We are sending uevent and masking the VWARNA and VWARNB interrupt
as soon as it occurs. when unmasking VWARNA we reset the BCUDISA o/p line
if vsys has recovered and is above the VWARNA threshold.

Category: device enablement
Domain: EM-PMIC
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: I3ca08fab98e6c765a2421aa35d8aa9e5579c4435
Signed-off-by: Yashodhan Kurtadkar <yashodhan.kurtadkar@intel.com>
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>