Configure IRQ 51 to handle TCO warning interrupt
BZ: 122004

When the watchdog timer is about to expire (2.4 seconds left),
the PMC catches the SMI event, and toggles a virtual GPIO
that generates an interrupt event at OS level.
This interrupt is used to kickoff the watchdog a last time and
trigger all CPU kernel backtraces in order to log as much as
possible of the system state before the expiration of the
timer and following reset.

Change-Id: Ic01d95da4699352df11f1785154195ddbf10e78e
Signed-off-by: Florent Auger <florent.auger@intel.com>
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>