EM/pmic-i2c: Handle EACCES
BZ: 149298

Issue Description: With unsigned kernel, charger register access would be blocked.
pmic-i2c driver returns error in this scenario. The client has to handle the error
to identify the EACCES and EIO. This makes the client driver to be aware of the
unsigned and invalid battery scenarios. This is not good since the client is generic
and not platform dependant.

Solution: Return success if the IPC returns EACCES.

Category: device enablement
Domain: EM-PMIC-I2C
Origin: internal
Upstream-Candidate: no, needs rework

Change-Id: I28f78e27ee8a310dd949b57d2ff893ebaa543fb8
Signed-off-by: Jenny TC <jenny.tc@intel.com>