[Telephony/MCD] Manage power management
BZ: 113753

Grab a wakelock when an event that is registered by the client
is detected so that we prevent the system to go into suspend state

There's a chance that the platform might go to suspend when the
MCD catch a modem event and the userland (MMGR) takes too much time to
read the event. Once the event is caught, the userland takes a
wakelock if the client has MCD opened and is interested in
current state.
The release of the wakelock is done when the poll handler is
fired (ie: the client is getting the notification).
The wakelock is taken with a 10ms timeout as a failsafe in
case the userland is not able to read the event
10ms is the maximum timeout we allow the MMGR to take to
read the event. Beyond that point, we consider MMGR to be
malfunctioning and allow the platform to go into suspend state anyway.

Change-Id: I8a74a7832384ee020b835b90e04569211470ea91
Signed-off-by: Guillaume Ranquet <guillaumex.ranquet@intel.com>