[HSU] remove noisy log about Rx FIFO not empty
BZ: 85637

The Rx FIFO will always be not empty if the HSU is
woken from suspend up by a preamble.
This is the case with the bcm4752 GPS chip when tracking.
Change this log from 'dev_info' to 'dev_dbg' in
order to not fill up the log and reduce power consumption.
The 'cfg->preamble' could be tested and log only
if this is set, but I do not think it is worth the
added complexity.

I/KERNEL..Warning: 1 bytes are received in RX fifo after RTS active for 11 us
I/KERNEL..flush abnormal data in rx fifo

Change-Id: Icd60dfa713635ef0f34816e58f1f670289fd1f06
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>