[Telephony/eDLP_Driver]: UL data path: bundling of IP frames
BZ: 92059

In order to improve UL data throughput, UL IP frames are grouped
and concatenated into a HSI PDU while the device is busy to send
some data.

A PDU is sent as soon as it is ready and the HW is able to send it.
So bundling is not systematic and is done while PDU are stored in
waiting queue.

Change-Id: Ie61673155397f74e10c030fd419bf9dc2a22bec2
Signed-off-by: Julie Daragnes <julie.daragnes@intel.com>