[PORT FROM R43][Telephony/HSI-eDLP] Avoid race condition when updating credit
BZ: 153640

If there is a preemption between the credit availability check
and the credit update: dlp_hsi_controller_push is interrupted by
another one invoked in complete_callback context or a start_tx
workqueue, data can be sent whereas there is no more credit.
This is a protocol violation.

This race condition can be avoided by having the check and the update
atomic.

Change-Id: I12a00e7ba9352cddb05d911f114ba808d3fa554d
Orig-Change-Id: I0c86c63cf112bc21fff033288f8d52240854d694
Signed-off-by: Julie Daragnes <julie.daragnes@intel.com>