r69001-ts-i2c: Change the slot ID for each touch point
BZ: 155138

Previously, we used touch ID as slot ID to track multiple points in
multi-touch protocol B.
Although there is no requirement in the definition of Linux multi-touch
protocol B that slot ID should start from index 0, we found in Android
input service, it uses 0 as the first index of the slot array.
However, the touch ID of r69001 starts from 1. This kind of mismatch
causes the Android framework not able to track the tenth point. So
we reduce the touch ID before using it as slot ID to make it start
from index 0.

Change-Id: If65cfa382ced6e6ae110319a8ac59a8f1104beaf
Signed-off-by: Leo Yan <leo.yan@intel.com>