misc, emmc_ipanic: Header for eMMC panic partition
BZ: 153671

The eMMC panic partition header is written first and contains trace
data.

The eMMC header (1st block of panic partition) contains simple trace
information. Offsets of log, callstack, and gBuffer are now written
in the 2nd block (but chronologically after said sections). A /proc
entry for the header is also created after reboot so that it can be
accessed directly.

Change-Id: I9cbebc23d128a23391bc8cc5a194c566cf1724f0
Signed-off-by: Christophe Guerard <christophe.guerard@intel.com>