reboot_target_uefi: use OsIndications EFI variable for "dnx" target
BZ: 154008

It has been decided that the dnx mode will not be supported as a load
option like it should have been. Finally, it relies on a dedicated and
absolutely out of the specifications path using the
OsIndicationsSupported and OsIndications EFI variables. The idea is,
in the context of the user asking for the "dnx" mode, we check for the
fifth bit availability in the OsIndicationsSupported bitmask EFI
variable and then we set the fifth bit in OsIndications bitmask EFI
variable accordingly. This fifth bit is named
EFI_OS_INDICATION_RESCUE_MODE and it indicates to the BIOS that we
request it to break the usual boot flow and go into Fastboot mode
instead.

Moreover, this patch adds a carriage return character to the error
strings which were missing it.

Change-Id: Ia2ddb1b1fb3981e32a4064d5bc533cb92f6c9a25
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>