i2c/dw_i2c: implement arbitration for shared host controller
BZ: 162832

For platforms where an i2c host is shared between multiple entities like a CPU,
and a u-controller, we need some synchronization mechanism for i2c access. The
actual synchronization is platform dependent.

This patch adds the support to introduce sync for such hosts. Also, we disable
the power management (also runtime) for such hosts to avoid lockup issues. We
currently use the flag "shared_host" to signify a shared host and implement the
synchronization.

TODO: Update the driver for UEFI changes

Change-Id: I740c3a446fdb42690c964ca057b77d533a3f9683
Signed-off-by: Sundar Iyer <sundar.iyer@intel.com>
Signed-off-by: Souvik Kumar Chakravarty <souvik.k.chakravarty@intel.com>
Signed-off-by: Alexander A Oganezov <alexander.a.oganezov@intel.com>
Signed-off-by: Bin Yang <bin.yang@intel.com>