drivers/misc: add rawio driver framework and sub-drivers to replace intel_soc_dump
BZ: 158045

Many people from MCG and other groups are using the debugfs based
SoCregister dump interface to debug android os and drivers and for
other purpose. The source file is in
arch/x86/platform/intel-mid/intel_soc_dump.c.

The design hardcoded register base address depending on different SoC.
This makes it hard to be scalable and for each new SoC we need modify
the driver. Now we have a new driver which is scalable and supposed
to work on all Intel Atom SoC without the need of change the driver
on supporing new SoCs.

We tried pushing these new drivers to LKML and got negative feedback.
The reason is that they prefer to do that from user space plus uio.
But we have special requirement that for some situation, e.g.
pre-silicon and early power-on days, it's inconvenient to integrate
any tools to rootfs, instead, a minimal kernel plus a minimal existed
rootfs is used, this limits us to implement it from userspace. But the
debugfs from kernel is always available in whatever situation so we
still need to manintain these drivers by ownselves in MCG(or even Intel).

Change-Id: I334a7da7be664a93715adf9dbfcb80a05a5ab002
Signed-off-by: Bin Gao <bin.gao@intel.com>