Touch screen: Filter out the invalid touch points
BZ: 155513

According to the datasheet of the touch panel, the touch id of the
point should be between 1 to 15. However, on some phones, invalid
point with a touch ID valued zero may appear when the touch panel
is just power on. Moreover, when the invalid point appears, the point
with touch ID 1 also becomes invalid. So we filter out these
invalid points as a workaround before we get a perfect solution.

Change-Id: I3e845ac85cc0d287d248c52bc2942f61518ed231
Signed-off-by: Leo Yan <leo.yan@intel.com>