reboot_target_uefi: don't use the BootNext variable anymore
BZ: 160170

It has been decided that following the standard is not the correct way
to go. This patch replaces the use of the BootNext variable and the
UEFI load options to influence the boot flow in favor of a brand new
and absolutely out of the specification variable nammed
LoaderEntryOneShot associated to an OSLoader dedicated GUID. The
target string is saved into this new variable, encoded as a CHAR16
string, and before reset is performed.

Change-Id: Ibb465728b8ff29adfc3321963a7ff524fec8e379