Revert "Configure IRQ 51 to handle TCO warning interrupt"
This reverts commit 5236a9d64426f9423678633202e30315822ad3f4.

This commit breaks the watchdog functionality. Because of some
changes missing in BIOS, TCO warnig interrupt gets triggered and we
see kernel panic during boot.On TCO warning interrupt will dump the
all CPU info for debug purpose and this will not break the watchdog
functionality on removal of this feature.
Reverting as this commit is only for debug purpose.

Change-Id: I25c4d63da23941a719609e955ceef2683005c1df
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>
Signed-off-by: Mantesh sarasetti <manteshx.sarasetti@intel.com>