emmc_ipanic: use partition label instead of partition device name
BZ: 162336

Emmc_ipanic driver currently use device name and partition number to
bind to the panic partition. When partition schema change, the
partition number need to be updated otherwise it could lead to
corrupted data on partitions, which as occurred in the past.

To prevent such issue, partition label could be used, it also better
scale when partitioning change.

Change-Id: Ic724dd8a35626241bbba710263b640bc68f7714e
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>