pmic: Fix null pointer access in crystal cove pmic
The driver has external APIs accessing internal driver
structures which might not be initialized if the device
probe is not called, in which case it results in NULL
pointer exception.
Access i2c APIs only if the probe happened and pointer
is valid.

Also couple of APIs (setb/clearb) are called from thermal
and EC drivers but are not external. If those drivers
are modules it will result in linker errors.

Change-Id: Ibbe87c6590f519b55c3b7d4e21eefff33bc65759
Signed-off-by: Huzefa Kankroliwala <huzefa.nomanx.kankroliwala@intel.com>