[Telephony/HSI]: HSI TX timeout due to illegal PDU state change
BZ: 161743

PDU status can be changed during a transfer, causing a HSI TX timeout.
PDU status is updated in both tx_complete callback and
dlp_net_start_xmit. The PDU to be sent is popped from waiting queue by
the tx complete callback and set to PROCEEDING state; in the meantime,
start_xmit calls dlp_send_pdu for the same PDU and set it back to
COMPLETED whereas it is not in waiting queue anymore and already
transitioned to PROCEEDING state. The PDU transfer aborts and  leads to
a HSI TX timeout.

dlp_net_send_pdu should check PDU status is PENDING before setting it to
COMPLETED.

Change-Id: I4cdb5d68ad63abae97e680b23473a0242c27a761
Signed-off-by: Julie Daragnes <julie.daragnes@intel.com>