intel_mid_gps: use acpi_get_gpio_by_name() for gpio numbers
BZ: 168246

We use acpi_get_gpio_by_name() to retrieve gpio numbers.
It implies that, for each GPIO that we need on the platform, the device
description in the ACPI table contains named integer constants with the
following names:

- ENAB for enable gpio
- RSET for reset gpio
- HSTW for hostwake gpio

the value of the variable must be the GPIO index in the resource description
list of the device.

example:
Name (ENAB, One) //if we need the enable GPIO and its index is 1.

Change-Id: I9cc14914b4b85e3b4b8af8a890c131247dce6aa4
Signed-off-by: Guillaume Aubertin <guillaume.aubertin@intel.com>
Signed-off-by: Nicolas SUET <nicolas.suet@intel.com>