iTCO_driver: Fix suspend/resume issue from S3
After resuming from S3 the NO_REBOOT bit of PMC_CFG being
set and resulting in failure to reboot after watchdog timeout.
Fix is to clear the NO_REBOOT bit after resume.

Also need to set the heartbeat value of watchdog, so that
it does not trigger the watchdog reset immediately after the
resume.

Change-Id: Ifef106d674ee515a94f35db58929c599b85f934a
Signed-off-by: Mantesh sarasetti <manteshx.sarasetti@intel.com>