scu log: Support for SCU Recoverable Fabric Errors
BZ: 174709

When the SCU sends a recoverable fabric error, it raises an IRQ.
This IRQ checks whether there is a pending recoverable FE or an
unsolicited SCU message.
Any incoming recoverable FE is then sent through TP2E, if enabled.
A stack ensures no recoverable is lost during back-to-back IRQs.

Change-Id: I1ea61c39194c20e7cd94dad1419bc97a623163be
Signed-off-by: Flavien Boutoille <flavienx.boutoille@intel.com>