OSNIB - move the OEM region to a free CMOS area
BZ: 152973

For legacy reason, an area in the middle of the OEM region was
found to be locked by the BIOS. It used to contain BIOS password.
To align with BIOS requirement that area must remain unaccessible
to the OS. This region is now skipped by adding a reserved area.
Current implementation of the OSNIB goes from 0x10 to 0x4F (64 bytes):
- Intel OSNIB : 0x10 to 0x2F
- OEM OSNIB : 0x30 to 0x4F with the locked section from 0x38 to 0x3F.

New implementation is (80 bytes):
- Intel OSNIB : 0x10 to 0x2F
- BIOS reserved: 0x30 to 0x3F
- OEM OSNIB : 0x40 to 0x5F.

Change-Id: I357e693e620406910c2f05e7887a6b578044f43c
Signed-off-by: Florent Auger <florent.auger@intel.com>