iTCO_wdt: Logging the wdt_keepalive()
BZ: 173915

Currently when system happening the wdt case, sometimes
it is due the tick count is inaccurate, sometimes it is
due watchdogd can not be executed in time.

Here we add one logging for wdt_keepalive() calling, it
will be very helpful for us to analysis the wdt issues,
and for scu wdt, there is already similar log there.

The output is like below:
[   11.503182] iTCO_wdt: iTCO_wdt_keepalive

Change-Id: I5aabd379f63dcd670b4eb70f6bf7ec720c152da6
Signed-off-by: Chuansheng Liu <chuansheng.liu@intel.com>