EM/bq24261: Handling charger fault recovery
BZ: 169326

Issue description: Fault recovery interrupt is not forwarded to IA
if charging or charger is disabled. The situation can occur, if a
throttling happens when the chip is in fault. Issue seen under this
scenario is that the charging does not resume when the chip recovers
from fault inspite of de-throttling.

Solution: Start a thread while the chip is in fault and when the
charger is connected. Montior the chip state for recovery on specific
intervals so that, the moment the chip has recovered, the health update
happens and charging can resume.

Change-Id: I6f53edc26c26e05430143088cdc8e96a5d92a418
Category: device enablement
Domain: BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, needs rework
Signed-off-by: ysiyer <yegnesh.s.iyer@intel.com>