vibra: Confirm that auto calibration is done before skipping the calibration
BZ: 178889

SCU configures the Vibra 2605 and puts it in STANDBY mode.
The driver gets the vibra out of standby mode first and then checks
if auto-calibration is already done to skip the calibration.

Since the register values for verifying the calibrations
are same in default and standby mode scenario, compare the
auto calibration compensation result register value with its
default value to confirm if calibration was actually done earlier.

Change-Id: I9480ebea5730924bfd9d70da6796b4797fbc574a
Signed-off-by: Samreen Nilofer <samreen.nilofer@intel.com>