HSU: protect from uninitialized hsu_port_func_cfg pointer
BZ: 179208

At least on CHT-RVP board w/A1 SoC, EDKII FW v36.76, there's a
kernel panic happening, similar to the one decribed in the Bugzilla
mentioned above.

The problem is that the null pointer dereference protection logic is
incorrectly implemented.

The port can be any unsigned integer in theory, but if for some
reason the hsu_port_func_cfg is NULL, and port is bigger than 0,
the "if (!cfg)" protection won't kick in, as the sum is not really
NULL.

So, the correct way to protect this will be to check hsu_port_func_cfg
first for NULL, and if it is, fail with -EINVAL, else do the operation
to determine cfg address.

Change-Id: I77308e6100e8b2ecf5cd764ceb1347a5a1220322
Signed-off-by: Sergio Aguirre <sergio.a.aguirre.rodriguez@intel.com>
Signed-off-by: Wang Yiyang <yiyang.wang@intel.com>