Pass parameter of timestamp sync to sensorhub firmware
BZ: 183443

Once, the psh firmware not send timestamp to linux, and the
timestamp added by hal. The sensorhubd receive the sensor report,
then pass to hal by socket, but the hal and sensorhubd is not
the same process, so sometime the timestamp added by hal is too late,
but sometime this timestamp is not too late, the system load
determines whether is too late or not too late. It means the
interval of timestamp is not stable, so we need psh insert timestamp
directly. To sync to psh timestamp with linux, we need pass linux
timstamp to psh.

Change-Id: Id631f3dcd6247d04e6b8ca1001696b47ccb43a7d
Signed-off-by: huijiang <hui.a.jiang@intel.com>