pwm/pwm_byt_core: Update bit is taking more time to get cleared.
BZ: 180221

Every set backlight call waits for update bit to get cleared
after writing new values, to be assure that values are updated,
This update bit is taking more time to be cleared.
In this patch we are not waiting for update bit to get cleared,
but before writing any new values we  check if update bit is
cleared which will mean last update was success or no update at
all and we can go ahead with new values. this implementation
not only saves the waiting time and makes sure before writing
new value update bit is cleared.

Change-Id: Ia6d4711c8dd27467228e0f3d26d185b6b3778c10
Signed-off-by: vkorjani <vikas.korjani@intel.com>