In MERR update cmdline offset location in signed OS
BZ: 186548

In signed OS the offset where kernel command line is read
is not the same as for unsigned images.

On Merrifield, with the new signing OS method the offset
has been increased. Now the command line must be read at the
address 0x000004D8.
For unsigned OS the command line pointer does not change.

Change-Id: Iaf2a091904a75e4b4c20f52b46694e272d6ebeb7
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>