[Telephony/HSI]: Secure READ function when driver removal
BZ: 206349

When the dlp flash driver is removed but the read function
is still in progress, the context structure is cleaned and
freed by the system that calls the cleanup function. Then
the read function is not properly stopped and handle
deallocated data. The patch stopped properly the reading
thread and check the data pointer before handling them.

Change-Id: Ie4ad4e33f5f0b21d3623393f85ec15e3c9f123bb
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>