[Telephony/HSI]: change order in tty_close function
BZ: 206349

In order to secure the tty close ans split the driver
aspect from the tty, the order of the commands in the
function has been changed: firstly the function handles
the tty data to ensure the clean close of the port then
the function manages the driver data, depending on
the driver state. If the dlp driver is removing, the tty
can close properly its data even the driver is already
removed by the system.

Change-Id: I866c94a8a676b1f03a3b66c0961d96ba86272cbb
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>