[Telephony/HSI]: change cmd tx error management
BZ: 206349

When the CMD sending fails, for the TX part, the
actions to be done are not the same regarding the
kind of error.
There are 3 different errors that imply different
actions because of the state of the sending process:
- If the TX sending function call fails, the message
pdu can be freed, because nothing more will be done.
- If the TX timeout happens, the cmd message is already
in the sending process pipe and we don't know where
the pdu is (still in the controller or already send
but too late). So the PDU can't be freed, but we can
avoid the call of the complete call back that is then
useless.
- If the error is due to the command status, then this
means the complete sending process has been executed
and the PDU will be freed, so nothing more to do is
needed.

The issue happens in this last case, in the TX path,
where the MSG context was forced to NULL but it was
still used by the complete callback function.

Change-Id: I38b36ea9e16671bb96a844b43cc9c9e612a211d4
Signed-off-by: Gwenn Bourree <gwenn.bourree@intel.com>