[Telephony/HSI-eDLP] fix 'count leak' of PDUs in control queue
BZ: 206349

When PDUs are not pushed to the controller due to the link being
in a bad state, an error is returned to the callee but the state
variables are still updated as if the message was sent (i.e.
ctrl_len / room are still set as if the message was actually
queued).

Fix the issue by factorizing the error handling for all cases in
a single code block.

Change-Id: I8ae208d111659eaf122b3a03070b6fd8681d6ddd
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>