EM/bq24261_charger: Reset charging before enabling charging
BZ: 204837

Enable charging may not succeed if chip is just out of a fault,
or if the pmic bat-zone is in a indeterminate state.

Hence, reset charging before enabling new charge-cycle again.
In case when the battery-temperature is within hysteresis
range around a zone boundary, have a fail-safe mechanism of writing
cc and cv from the bq24261-driver as well.

Change-Id: I70103b8173003f234be1d2f078abbe6426e85750
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>