[Telephony/HSI-eDLP] do not hog CPU with bad signature messages
BZ: 206349

As the DLP Trace client displays a hex dump of the message
header each time it detects one with a bad signature, it can
quickly use all the resources of a CPU due to the vast amount
of logging this entails.

Fix the issue by only logging bad signature errors on the first
detected packet and by tracking the actual number of invalid
logging packets.

Change-Id: I4adf4d0426ef71b5c50f19f97fb048cc4143a320
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>