[Telephony/HSI-eDLP] Fix error returned by dlp_flash_dev_write
BZ: 206349

In case the phone is rebooted while the modem is being flashed,
the DLP flash driver returns a wrong error message to the flash
client leading the latter to enter an 'attempt to write' loop
(as the error is not a 'fatal' one). This in turn lead to a
software watchdog at kernel side.

Fix the issue by returning a proper error code when the error is
'definitive' (i.e. when a retry will not fix it).

Change-Id: I0f11ecc4a09d24baff28d9252783efead0f90de9
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>